move grid definitions factories under Definition namespace
