+++ src/Core/Grid/Definition/Factory/GridDefinitionFactoryInterface.php
@@ -24,7 +24,7 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShop\PrestaShop\Core\Grid\Factory\GridDefinition;
+namespace PrestaShop\PrestaShop\Core\Grid\Definition\Factory;
 
 use PrestaShop\PrestaShop\Core\Grid\Definition\GridDefinitionInterface;
